pod lib lint

+++ Classes/NSString+LevenshteinDistance.m
@@ -6,8 +6,8 @@
 
 - (NSUInteger)levenshteinDistanceTo:(NSString *)otherString
 {
-    int             height   = self.length + 1;
-    int             width    = otherString.length + 1;
+    NSUInteger height = self.length + 1;
+    NSUInteger width = otherString.length + 1;
     NSMutableArray *distance = [[NSMutableArray alloc] initWithCapacity:width * height];
     NSUInteger      x, y, cost;
 
@@ -30,9 +30,9 @@
             unichar c2 = [otherString characterAtIndex:x - 1];
             cost = c1 == c2 ? 0 : 1;
 
-            int insert  = ((NSNumber *) distance[(y - 1) * width + x]).integerValue + 1;
-            int remove  = ((NSNumber *) distance[y * width + x - 1]).integerValue + 1;
-            int replace = ((NSNumber *) distance[(y - 1) * width + x - 1]).integerValue + cost;
+            NSUInteger insert  = ((NSNumber *) distance[(y - 1) * width + x]).integerValue + 1;
+            NSUInteger remove  = ((NSNumber *) distance[y * width + x - 1]).integerValue + 1;
+            NSUInteger replace = ((NSNumber *) distance[(y - 1) * width + x - 1]).integerValue + cost;
             distance[y * width + x] = @(min(min(insert, remove), replace) );
 //            NSLog(@"d(%d,%d) = %@", y, x, distance[y * width + x]);
         }
